Changes to Issue #3 , iPod like scrollwheel
Includes an interpreter to get the amount of water that should be send
to the arduino. The amount of water comes from the ipod.html and ist
postsend to the water.php and further to the waterbar.php.
Changes the height of the waterbar to fit the small iframe in the
watering popup.
Some changes in the texts.
Some cleanup.