Minor changes in database query to fix Issue #5
Displayed messages of the plant reduced to 3. That until now wouldn't
have any effect because this messagecounter was overwritten in the
index.php. But that is fixed, too.
Grab the amount of water that was given to the soil in dependency to the
time when the watering event was recognized and a message was sent out.