Adress Issues #33 #35 #42
Only try 5 times to connect to the website with the database, then go on
with the moisture measurement.
Don't send the RAM values to the database.
Use the local IP adress of the website with the database instead of
using DNS. That makes the upload process more stable and faster.